hw/crypto/xlnx-asu-aes-new: introduce a new version of the ASU AES model
This commit adds a skeleton for a new ASU AES model. This model will
focus on feature completeness while avoiding dependency to external
code. The ASU keyvault is not included in this model. This model will
rely on a known interface with the keyvault to perform required
operations (key loading, key decryption and zeroization).

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>
[FK: 260324 - fixed class_init() to take a const data argument
              (12d1a768bdf qom: Have class_init() take a const data argument)
            - fixed against property constification and DEFINE_PROP_END_OF_LIST
              removal (5fcabe628b8 include/hw/qdev-properties: Remove
                       DEFINE_PROP_END_OF_LIST)
     260304 - fixed against reset API changes (ad80e367447 hw, target: Add
              ResetType argument to hold and exit phase methods)]
Signed-off-by: Frederic Konrad <fkonrad@amd.com>