hw/crypto/xlnx-asu-aes-new: implement IV/INTMAC/S0 load operations
Implement the IV, INTMAC and S0 load operations in the ASU AES model.
Those operations copy the value given in the corresponding x_IN_y
registers in the internal state of the model.

When countermeasures are enabled, the mask of the object being loaded is
xor'ed with it to form the final data.

Note that the model use one block to store both the GCMLEN and the S0
data. GCMLEN is used when performing AES-GCM encryption/decryption,
while S0 is used in the case of AES-CCM. Those operations are mutually
exclusive so the same block can be reused internally.

Also note that the GCMLEN_IN data is loaded on a INTMAC load operation,
alongside the INT_MAC_IN data. This is because a INTMAC load operation
is used when restoring the context of a AES-GCM operation. INTMAC and
GCMLEN are the two AES-GCM specific datum to be restored in this case.
The controller does not expose a specific operation to load GCMLEN. The
internal AES engine IP used in this controller does not specify when it
effectively load this data, but specifies that it should be set before
the corresponding INTMAC loading pin is asserted (alongside INTMAC
data).

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>