hw/crypto/xlnx-asu-aes-new: implement the soft reset mechanism
Implement the soft reset mechanism in the ASU AES model. After a hard
reset, the device is in soft reset state, from which it can be taken out
by writing to the SOFT_RST register.

While in soft reset, the device clear the (not yet implemented) input
FIFO and zeroize the AES engine. It also advertises itself as busy and
not ready in the status register.

Note that we directly wire the busy flag in the STATUS register to the
soft reset state. Apart from being set when in soft reset, this flag is
normally used to signal that the AES engine is currently performing an
operation and has not produced the corresponding data yet. Since the AES
engine will be modelled in a way that all operations are performed
atomically, the guest cannot possibly observe this flag as being set,
other than the case where the device is in soft reset.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>