hw/crypto/xlnx-asu-aes-new: implement the AES-ECB mode
Implement the AES-ECB cipher mode in the ASU AES device. The result is
stored in the out internal buffer, ready to be sent in the sink stream.
The sending operation will come with next commits.

The ECB mode is as simple as it can get. Each encrypted or decrypted
block is independent from the previous one. No intermediate state is
needed.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>