hw/crypto/xlnx-asu-aes-new: implement the AES-CCM mode
Implement the AES-CCM mode in the ASU AES model. Additionally to the
encryption process, this mode produces a MAC (or tag) at the end of the
process, used to verify the authenticity of the cipher and the AAD (the
additional authenticated data) that can be optionally provided before
the data.

This introduces the notion of authentication phase. When the AUTH flag
is set in the MODE_CONFIG register, the device processes the received
data as authentication data. This must happen at the beginning of the
cipher process. When no more authentication data are to be sent, this
flag is cleared by the guest and the following data are then handled as
text.

The AES-CCM mode is a combination of an AES-CTR for the ciphering part,
and a CBC-MAC for the authentication part. The only difference with
AES-CTR is that the initial IV value is not used to produce the first
cipher text block. Instead, it is used to produce the S0 value (see
below).

To build the MAC, the AES-CCM mode performs consecutive xors:
   - It starts by encrypting the first AAD block
   - The next AAD block is then xor'ed with the result of the previous
     operation, and the xor result is in turn encrypted.
   - This is repeated for all AAD blocks
   - When the authentication phase ends and the controller start
     receiving the text blocks, the plain text is used in place of the
     AAD block, the process remains the same, xor with previous result
     and encryption.
   - At the end of the whole cipher process, the MAC must be finalized.
     It is xor'ed with the S0 value, and encrypted one last time to
     produce the final MAC value. The S0 value is the result of the
     encryption of the initial IV value.

The MAC finalization is triggered when EOP is received:
   * in non-authentication phase,
   * or when in authentication phase and when the flag
     AUTH_WITH_NO_PAYLOAD is set in MODE_CONFIG.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>