hw/crypto/xlnx-asu-aes-new: implement the AES-GCM mode
Implement the AES-GCM mode in the ASU AES model. This mode produces a
the cipher data and a MAC based on authentication data and the cipher
text.

AES-GCM is a combination of an AES-CTR for the cipher part, and a GMAC
for the MAC part. GMAC relies on the GHASH function to perform successive
MAC computations. The GHASH function can be defined as:
   GHASH_k(a, b) = (a xor b) . E_k

With k the key used for the ciphering operation, and E_k the results of
the encryption of a 128 bits block full of 0s with the key k. The "."
operator represents the multiplication in GF(2^128) with the field
polynomial 1 + x + x^2 + x^7 + x^128.

The MAC finalization consists in:
   - a GHASH between the MAC and the "len(A) || len(C)" value. This
     value is the concatenation of the size of the AAD and the size of
     the cipher text, both represented with 64 bits integers (forming a
     128 bits block).
   - a final xor with the encrypted J0, J0 being the first counter (IV)
     value.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>