hw/crypto/xlnx-asu-aes-new: implement the AES-CMAC mode
Implement the AES-CMAC mode in the ASU AES model. This mode is
authentication only, it does not produce cipher text. It is similar to
CBC-MAC (used in AES-CCM) as it only differs for the finalization
operation. The last text block is xor'ed with a derivation of the key
before being normally processed.

From an implementation point of view, we handle this operation as a
classical streamed operation, but without sending the content of the
output buffer at the end of each step.

The input buffer padding is done with 10....0 instead of zeros only.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>