hw/crypto/xlnx-asu-aes-new: implement the key decryption operation
Implement the key decryption operation in the ASU AES model. This
operation is used to decrypt a black key from the vault, and send back
the corresponding red key.

This operation does not use the classical receive/send data path and
bypass the FIFOs. It directly uses the AES engine wrapped in the device.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>