hw/misc/xilinx-psxc-rpu-cluster-core: add a skeleton
Add a skeleton of the PSXC RPU cluster core peripheral. This model is
meant to merge and replace the four models:
  - xlnx-versal-net-rpu-cluster-core0.c
  - xlnx-versal-net-rpu-cluster-core1.c
  - xlnx-psxc-rpu-cluster-core0.c
  - xlnx-psxc-rpu-cluster-core1.c

This skeleton includes all the GPIOs, with the same interface as the
models it aims at replacing. It also includes register decoding for both
versal-net and versal2. versal2 register map is taken as the base
map, and the model performs versal-net -> versal2 translation when the
version property is set to versal-net.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>
[FK: 260324 - fixed class_init() to take a const data argument
              (12d1a768bdf qom: Have class_init() take a const data argument)
            - fixed against property constification and DEFINE_PROP_END_OF_LIST
              removal (5fcabe628b8 include/hw/qdev-properties: Remove
                       DEFINE_PROP_END_OF_LIST)
     260304 - fixed against reset API changes (ad80e367447 hw, target: Add
              ResetType argument to hold and exit phase methods)]
Signed-off-by: Frederic Konrad <fkonrad@amd.com>