versal: ddrmc-xmpu: Support NPI addressing of DDRMC5 XMPU
The XMPU register offsets in the NPI address-space are different
between the original DDRMC XMPU and newer DDRMC5* variants.

This adds support for the differences.

To select DDRMC5* offsets, an object instance's "regs-variant"
string property should have a prefix of "ddrmc5".  Suffix is
optional and, if any, must not start with a digit.

Signed-off-by: Tong Ho <tong.ho@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>