etrace: don't use fast tb lookup
Previous attempt to disable fast lookup in the lookup_tb_ptr helper fails to
catch the case where tb_lookup generates an exception.  In this case
cpu_loop_exit is called long jump is taken and etrace_dump_exec_end is called
with a wrong PC (the one of the next tb).

Avoid and simplify this by setting the CF_NO_GOTO_PTR flags when exec trace are
enabled.

Signed-off-by: Frederic Konrad <fkonrad@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>