target/arm/tcg/op_helper: wfi: revert to vanilla implementation
Our for has diverged significantly from vanilla regarding the wfi
instruction implementation for ARM CPUs. Initially, this comes from the
need of having an implementation of the STANDBYWFI output signal. This
signal is set whenever the CPU is in wfi state. Upstream does not have
this signal.

The current state of the wfi instruction is not ideal:
   - it disables the optimization where if the CPU has work, the
     instruction returns immediately instead of going back to the main
     loop,
   - it never puts the CPU into halted mode. It generates a EXCP_YIELD
     exception instead, effectively turning the wfi instruction into a
     costly nop.
   - the STANDBYWFI implementation is actually broken. The signal is
     raised in HELPER(wfi) but never lowered. The lowering calls might
     have been lost in successive rebases.

This commit reverts the HELPER(wfi) function to the current vanilla
state. Next commits will reintroduce proper STANDBYWFI implementation.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>
[FK: 260423 - adapted to v10.2.0, basically reverting back to v10.2.0 instead of
              v8.2.7]
Signed-off-by: Frederic Konrad <fkonrad@amd.com>