target/arm: reintroduce the STANDBYWFI implementation
Reintroduce the STANDBYWFI signal implementation for ARM CPUs. This
signal is raised when the CPU enters wfi state. It is then
unconditionally cleared in arm_cpu_exec_interrupt since the only way for
a CPU to wakeup from wfi in QEMU is to receive an interrupt.

Regarding the optimization in HELPER(wfi) where the instruction is
skipped if cpu_has_work returns true, the signal is simply pulsed. This
will allow peers to still record the wfi event.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>