hw/misc/xlnx-versal-apu-ctrl: remove the flush request check
The pwrctl_postw function was checking whether the CPU was in wfi state
before allowing the flush request. It was relying on the is_in_wfi
internal CPU state. This variable is not in used anymore and is about to
be removed.

Remove the check altogether. As the comment suggested, it is not ideal
to rely on CPU internal state here. In the unlikely case where this
functionality is actually relied upon, it could be added back by wiring
the wfi CPU signal to this device model.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>