hw/gpio/xilinx_iomod_gpi: allow input sampling
Add the gpi-sample-mask property to allow input sampling. When a bit is
set in the mask, the corresponding input is sampled on a rising edge.
When reading the GPI register, the guest will read the sampled value
or'ed with the current input value, and the sampled data will be
cleared.

This behaviour does not exist on real hardware and is a workaround to
allow the ZynqMP platform to correctly works with the vanilla QEMU
implementation of the wfi instruction. In the power management protocol
between the ARM cores and the PMU, when an ARM core requests to go power
off, the PMU will wait for it to enter wfi mode by (indirectly) watching
the STANDBYWFI signal. The flow is as follows:

       [ APU core ]-STANDBYWFI-> [ combinatorial logic ] -
                                                          |
  [ PMU core ] <- [ iomodule INTC ] <- [ iomodule GPI ] <-

When receiving an GPI IRQ, the PMU core will check the GPI register to
know what core entered wfi state. It is totally possible that the PMU
actually misses the STANDBYWFI signal because the GPI does not sample
it. In that case the PMU will simply ignore the interrupt and the APU
core will be stuck.

This can happen if the APU core receives an IRQ before the PMU has time
to check the GPI register. In that case the APU will wake up from wfi
and the STANDBYWFI signal will go low.

This is a flaw in the ZynqMP design: not sampling the STANDBYWFI signal
is really an issue in this case. Whether the whole thing works or not is
highly dependent of the ARM CPU implementation (the ARM ARM is quite
permissive regarding WFI implementation). This hack allows to keep the
QEMU vanilla wfi instruction implementation and not penalize other
platforms.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>