hw/core/remote-port: replace the Chardev with a QIOChannel
Replace the Chardev in the remote port device with a QIOChannelSocket.
A QIOChannel is more suited to the needs of the remote port:
   - It is more low-level than a Chardev. The Chardev is not really
     necessary here because we only need to exchange data with the peer.
   - It is a modern mechanism with support for complex operations such
     as read/write_all variants of the transfer functions,
     blocking/non-blocking option and coroutine support. This will come
     very handy for the upcoming remote port rework.

It comes as a compromise regarding backward-compatibility. Until now the
remote port was explicitly exposing the Chardev to the user, through
various means:
   1. The `chardev' qdev link property allowed to directly pass a
      Chardev to the remote port,
   2. The `chrdev-id' qdev string property was the same idea as 1., but
      passing the Chardev ID instead of the object itself,
   3. The `chardesc' qdev string property could be used to describe the
      Chardev to be created by the remote port during the realize phase.
      The syntax is the one used by standard Chardev creation on the
      command line.
   4. Implicit Chardev creation if all the above failed, by creating a
      UNIX socket in the machine path directory

Keeping backward compatibility with methods 1. and 2. is not possible,
because this commit completely removes the Chardev usage in the remote
port code.

This commit keeps backward compatibility with method 3. though, by
locally parsing the chardesc string and creating the corresponding
QIOChannelSocket connection. Only socket related options are accepted,
other options will raise an error.

Method 4. is kept as well without behaviour change.

This seems a reasonable compromise. The most used method by far is 4.
Method 3. comes next, usually through the -global command line argument.
Method 2. only occurrence is in the hw/riscv/virt.c file. Corresponding
usage could not be found in the various testsuites. Finally Method 1.
usage could not be found. If this commit breaks some existing command
lines or hwdtb, it should be easy to fix them by converting them to use
method 3.

Signed-off-by: Luc Michel <luc.michel@amd.com>