hw/core/remote-port: replace rp_reset with rp_machine_done
rp_reset is currently creating the remote port thread, effectively
starting communication with the peer. What we want here is to ensure
that the machine is fully created and reset before starting
communication and handling peer packets. This ensures peer packets will
be handled after our local reset. This is important to not overwrite
peer writes with local reset writes.

Use a notifier on the main loop instead. When the main loop is started,
we are sure that the initial reset is done.

Signed-off-by: Luc Michel <luc.michel@amd.com>