hw/core/remote-port: rework the receiving logic
This commit reworks the receiving logic of the remote port.

In its current state, the remote port spawns a thread to handle the
received packed on the iochannel (the "protocol thread"). A received
packet can follows different paths:
   - If it is a response to a previously sent packet, the protocol
     thread will handle it inline and write it in appropriate structure.
     The progress_cond condition is then signaled for another thread
     waiting for the packet to make progress.
   - Otherwise, the packet is forwarded (rp_pt_handover_pkt) to the main
     QEMU thread for handling. It is written into the RX queue and the
     main QEMU thread is signaled using an event mechanism based on
     pipes.

This code has aged and does not leverage modern async I/O facilities
offered by QEMU. Moreover this ping-pong between threads leads to a lot
of thread synchronisation hitting the performances.

This commit reworks it the following way:
   - The protocol thread is replaced by a QEMU coroutine. This coroutine
     is in charge of receiving the asynchronous packets. The QIOChannel
     is set non-blocking. This has the direct consequence that the
     coroutine will yield when waiting for a packet using the
     qio_channel_.*_all family functions. It will be automatically
     re-entered by the QEMU main loop when data arrive on the iochannel.

   - The introduced `recv_one' function waits for the reception of one
     packet. It is safe to be called from any context. The coroutine
     calls it in a loop to handle the asynchronous packets as mentioned
     above. Since the coroutine executes on the main loop, it holds the
     BQL so the packet can be handled inline without concurrency issue.
     Devices (vCPU threads, or the main thread during e.g., initial
     reset) sending packets and waiting for a synchronous response also
     call `recv_one'. Devices also hold the BQL since the remote port is
     a device. If the request is done by a vCPU thread, instead of
     relying on the coroutine to receive the response, it directly
     waits on the iochannel using qio_channel_wait, while holding the
     BQL. This allows inline receiving of the response and again avoids
     thread synchronisation.

This results in performance improvement and code simplification. The
performance improvement comes from the (almost) complete removing of
thread synchronisation:
  - the coroutine can now handle asynchronous packets directly,
  - the vCPU threads can send a request and synchronously wait for the
    response.

Note that the rsp_mutex and write_mutex are completely removed, due to
the fact that we now completely rely on the BQL. Other various
synchronisation objects are not used anymore and removed by this commit.
The RX queue is removed as well (the coroutine both receive and handle
the asynchronous packet).

This also reverts 3d62c71e4d81. This new flow guarantees that
synchronous responses and side-effects following a request to the remote
peer are handled in-order.

Signed-off-by: Luc Michel <luc.michel@amd.com>